Support log-in via Google SSO
We add a new Login page, and move overview logic into a separate
component.  The login page and ProfileButton use the react-google-login
module to get an access token that's exchanged with the backend.
Session tokens are then used to maintain login state.

We create an instance of axios so that the X-Requested-With default
header can be added, which is used by the backend to protect against
CSRF.

This change requires backend support.